New CRV API (#32)
* updated CRV API

* WIP new rand API

* WIP new rand API

* updated random name gen to only gen with normal characters

* fixed really dum typo

* Fixed all crv api bugs

* Update README.md

+++ src/main/scala/chiselverify/crv/Randc.scala
@@ -8,17 +8,17 @@ trait Randc {
   /** Returns the current value of the variable
     * @return return the current value of the variable
     */
-  def value(): BigInt
+  def value(): Int
 
   /** Gets the next value of the random variable
     * @return return the next value of the variable
     */
-  def next(): BigInt
+  def next(): Int
 
   /** Set the value of the variable
     * @param that the value to be set
     */
-  def setVar(that: BigInt): Unit
+  def setVar(that: Int): Unit
 
   override def toString: String = value().toString
 }
